+++ Models/Models/Bid.cs
@@ -13,7 +13,7 @@ namespace Models
         {
             
         }
-        public Bid(int amount, double price, Currency currency) : base(amount, price, currency)
+        public Bid(double amount, double price, Currency currency) : base(amount, price, currency)
         {
         }
